Code review (warning and more)
I make sure to remove all the warning about the TJL stuff, and also
corrected fews things.

+++ src/cgame/etj_rotation_matrix.hpp
@@ -15,50 +15,47 @@ public:
 
 	void init()
 	{
-		_m11 = 0;_m12 = 0;_m13 = 0;_m14 = 0;
-		_m21 = 0;_m22 = 0;_m23 = 0;_m24 = 0;
-		_m31 = 0;_m32 = 0;_m33 = 0;_m34 = 0;
-		_degree = 0;
+		_m11 = 0.0f; _m12 = 0.0f; _m13 = 0.0f; _m14 = 0.0f;
+		_m21 = 0.0f; _m22 = 0.0f; _m23 = 0.0f; _m24 = 0.0f;
+		_m31 = 0.0f; _m32 = 0.0f; _m33 = 0.0f; _m34 = 0.0f;
+		_degree = 0.0f;
 	}
 
 	void rotateAlongAnAbitaryAxis(vec3_t center, vec3_t axis)
 	{
 		// Use centroid of the object of rotation origin.
-		double a = center[0];
-		double b = center[1];
-		double c = center[2];
+		float a = center[0];
+		float b = center[1];
+		float c = center[2];
 
 		// Use vector (0,0,1) as Z-Up-Vector for the rotation
-		double u = axis[0];
-		double v = axis[1];
-		double w = axis[2];
+		float u = axis[0];
+		float v = axis[1];
+		float w = axis[2];
 
 		// Set some intermediate values.
-		double u2 = u*u;
-		double v2 = v*v;
-		double w2 = w*w;
-		double cosT = std::cos(_degree);
-		double oneMinusCosT = 1 - cosT;
-		double sinT = std::sin(_degree);
-		
+		float u2 = u*u;
+		float v2 = v*v;
+		float w2 = w*w;
+		float cosT = std::cos(_degree);
+		float oneMinusCosT = 1.0f - cosT;
+		float sinT = std::sin(_degree);
+
 		// Build the matrix entries element by element.
 		_m11 = u2 + (v2 + w2) * cosT;
 		_m12 = u*v * oneMinusCosT - w*sinT;
 		_m13 = u*w * oneMinusCosT + v*sinT;
-		_m14 = (a*(v2 + w2) - u*(b*v + c*w))*oneMinusCosT
-			+ (b*w - c*v)*sinT;
+		_m14 = (a*(v2 + w2) - u*(b*v + c*w))*oneMinusCosT + (b*w - c*v)*sinT;
 
 		_m21 = u*v * oneMinusCosT + w*sinT;
 		_m22 = v2 + (u2 + w2) * cosT;
 		_m23 = v*w * oneMinusCosT - u*sinT;
-		_m24 = (b*(u2 + w2) - v*(a*u + c*w))*oneMinusCosT
-			+ (c*u - a*w)*sinT;
+		_m24 = (b*(u2 + w2) - v*(a*u + c*w))*oneMinusCosT + (c*u - a*w)*sinT;
 
 		_m31 = u*w * oneMinusCosT - v*sinT;
 		_m32 = v*w * oneMinusCosT + u*sinT;
 		_m33 = w2 + (u2 + v2) * cosT;
-		_m34 = (c*(u2 + v2) - w*(a*u + b*v))*oneMinusCosT
-			+ (a*v - b*u)*sinT;	
+		_m34 = (c*(u2 + v2) - w*(a*u + b*v))*oneMinusCosT + (a*v - b*u)*sinT;
 
 		return;
 	}
@@ -72,16 +69,16 @@ public:
 
 		return;
 	}
-	
-	double getDegree(){ return _degree; }
-	void setDegree(double d){ _degree = d; }
+
+	float getDegree() { return _degree; }
+	void setDegree(float d) { _degree = d; }
 
 private:
-	double _m11, _m12, _m13, _m14;
-	double _m21, _m22, _m23, _m24;
-	double _m31, _m32, _m33, _m34;
+	float _m11, _m12, _m13, _m14;
+	float _m21, _m22, _m23, _m24;
+	float _m31, _m32, _m33, _m34;
 
-	double _degree;
+	float _degree;
 
 };
 #endif
